Shresth/requirements.txt
# PyTorch and torchvision
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=9.0.0
rasterio>=1.3.0

# Data processing and visualization
numpy>=1.23.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Metrics and utilities
scikit-learn>=1.2.0
tqdm>=4.65.0

# TensorBoard for logging
tensorboard>=2.12.0

# XGBoost for ensemble
xgboost>=2.0.0
joblib>=1.3.0
